Add configurable `LogEngine` for less opinionated logging
