+++ core/api/core.api
@@ -148,38 +148,6 @@ public final class com/juul/kable/LazyDescriptor : com/juul/kable/Descriptor {
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class com/juul/kable/Logging {
-	public fun <init> ()V
-	public final fun data (Lkotlin/jvm/functions/Function1;)V
-	public final fun getFormat ()Lcom/juul/kable/Logging$Format;
-	public final fun getLevel ()Lcom/juul/kable/Logging$Level;
-	public final fun setFormat (Lcom/juul/kable/Logging$Format;)V
-	public final fun setLevel (Lcom/juul/kable/Logging$Level;)V
-}
-
-public final class com/juul/kable/Logging$Data {
-	public fun <init> ()V
-	public final fun getLowerCase ()Z
-	public final fun getSeparator ()Ljava/lang/String;
-	public final fun setLowerCase (Z)V
-	public final fun setSeparator (Ljava/lang/String;)V
-}
-
-public final class com/juul/kable/Logging$Format : java/lang/Enum {
-	public static final field Compact Lcom/juul/kable/Logging$Format;
-	public static final field Multiline Lcom/juul/kable/Logging$Format;
-	public static fun valueOf (Ljava/lang/String;)Lcom/juul/kable/Logging$Format;
-	public static fun values ()[Lcom/juul/kable/Logging$Format;
-}
-
-public final class com/juul/kable/Logging$Level : java/lang/Enum {
-	public static final field Data Lcom/juul/kable/Logging$Level;
-	public static final field Events Lcom/juul/kable/Logging$Level;
-	public static final field Warnings Lcom/juul/kable/Logging$Level;
-	public static fun valueOf (Ljava/lang/String;)Lcom/juul/kable/Logging$Level;
-	public static fun values ()[Lcom/juul/kable/Logging$Level;
-}
-
 public final class com/juul/kable/ManufacturerData {
 	public fun <init> (I[B)V
 	public final fun getCode ()I
@@ -259,6 +227,7 @@ public abstract interface class com/juul/kable/Scanner {
 
 public final class com/juul/kable/ScannerKt {
 	public static final fun Scanner (Ljava/util/List;)Lcom/juul/kable/Scanner;
+	public static final fun Scanner (Ljava/util/List;Lkotlin/jvm/functions/Function1;)Lcom/juul/kable/Scanner;
 	public static synthetic fun Scanner$default (Ljava/util/List;ILjava/lang/Object;)Lcom/juul/kable/Scanner;
 }
 
@@ -378,5 +347,58 @@ public final class com/juul/kable/WriteType : java/lang/Enum {
 	public static final field WithoutResponse Lcom/juul/kable/WriteType;
 	public static fun valueOf (Ljava/lang/String;)Lcom/juul/kable/WriteType;
 	public static fun values ()[Lcom/juul/kable/WriteType;
+}
+
+public abstract interface class com/juul/kable/logs/LogEngine {
+	public abstract fun assert (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public abstract fun debug (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public abstract fun error (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public abstract fun info (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public abstract fun verbose (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public abstract fun warn (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+}
+
+public final class com/juul/kable/logs/Logging {
+	public fun <init> ()V
+	public final fun data (Lkotlin/jvm/functions/Function1;)V
+	public final fun getEngine ()Lcom/juul/kable/logs/LogEngine;
+	public final fun getFormat ()Lcom/juul/kable/logs/Logging$Format;
+	public final fun getLevel ()Lcom/juul/kable/logs/Logging$Level;
+	public final fun setEngine (Lcom/juul/kable/logs/LogEngine;)V
+	public final fun setFormat (Lcom/juul/kable/logs/Logging$Format;)V
+	public final fun setLevel (Lcom/juul/kable/logs/Logging$Level;)V
+}
+
+public final class com/juul/kable/logs/Logging$Data {
+	public fun <init> ()V
+	public final fun getLowerCase ()Z
+	public final fun getSeparator ()Ljava/lang/String;
+	public final fun setLowerCase (Z)V
+	public final fun setSeparator (Ljava/lang/String;)V
+}
+
+public final class com/juul/kable/logs/Logging$Format : java/lang/Enum {
+	public static final field Compact Lcom/juul/kable/logs/Logging$Format;
+	public static final field Multiline Lcom/juul/kable/logs/Logging$Format;
+	public static fun valueOf (Ljava/lang/String;)Lcom/juul/kable/logs/Logging$Format;
+	public static fun values ()[Lcom/juul/kable/logs/Logging$Format;
+}
+
+public final class com/juul/kable/logs/Logging$Level : java/lang/Enum {
+	public static final field Data Lcom/juul/kable/logs/Logging$Level;
+	public static final field Events Lcom/juul/kable/logs/Logging$Level;
+	public static final field Warnings Lcom/juul/kable/logs/Logging$Level;
+	public static fun valueOf (Ljava/lang/String;)Lcom/juul/kable/logs/Logging$Level;
+	public static fun values ()[Lcom/juul/kable/logs/Logging$Level;
+}
+
+public final class com/juul/kable/logs/SystemLogEngine : com/juul/kable/logs/LogEngine {
+	public static final field INSTANCE Lcom/juul/kable/logs/SystemLogEngine;
+	public fun assert (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public fun debug (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public fun error (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public fun info (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public fun verbose (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
+	public fun warn (Ljava/lang/Throwable;Ljava/lang/String;Ljava/lang/String;)V
 }
