+++ core/src/androidMain/kotlin/BluetoothDevice.kt
@@ -13,6 +13,7 @@ import android.os.Build
 import android.os.Handler
 import android.os.HandlerThread
 import com.juul.kable.gatt.Callback
+import com.juul.kable.logs.Logging
 import kotlinx.coroutines.android.asCoroutineDispatcher
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.newSingleThreadContext
