+++ core/src/androidMain/kotlin/Observers.kt
@@ -1,6 +1,7 @@
 package com.juul.kable
 
-import android.util.Log
+import com.juul.kable.logs.Logger
+import com.juul.kable.logs.Logging
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.filter
@@ -49,7 +50,10 @@ internal sealed class AndroidObservationEvent {
  */
 internal class Observers(
     private val peripheral: AndroidPeripheral,
+    logging: Logging
 ) {
+
+    private val logger = Logger(logging, tag = "Kable/Observers")
 
     val characteristicChanges = MutableSharedFlow<AndroidObservationEvent>()
     private val observations = Observations()
@@ -82,7 +86,7 @@ internal class Observers(
                 } catch (e: NotReadyException) {
                     // Silently ignore as it is assumed that failure is due to connection drop, in which case Android
                     // will clear the notifications.
-                    Log.d(TAG, "Stop notification failure ignored.")
+                    logger.debug { message = "Stop notification failure ignored." }
                 }
             }
         }
