--- core/src/androidMain/kotlin/Tag.kt
@@ -1,3 +1,0 @@
-package com.juul.kable
-
-internal const val TAG = "Kable"
