+++ core/src/androidMain/kotlin/gatt/Callback.kt
@@ -9,10 +9,7 @@ import android.bluetooth.BluetoothProfile.STATE_CONNECTED
 import android.bluetooth.BluetoothProfile.STATE_CONNECTING
 import android.bluetooth.BluetoothProfile.STATE_DISCONNECTED
 import android.bluetooth.BluetoothProfile.STATE_DISCONNECTING
-import android.util.Log
 import com.juul.kable.ConnectionLostException
-import com.juul.kable.Logger
-import com.juul.kable.Logging
 import com.juul.kable.State
 import com.juul.kable.State.Disconnected.Status.Cancelled
 import com.juul.kable.State.Disconnected.Status.CentralDisconnected
@@ -22,8 +19,6 @@ import com.juul.kable.State.Disconnected.Status.LinkManagerProtocolTimeout
 import com.juul.kable.State.Disconnected.Status.PeripheralDisconnected
 import com.juul.kable.State.Disconnected.Status.Timeout
 import com.juul.kable.State.Disconnected.Status.Unknown
-import com.juul.kable.TAG
-import com.juul.kable.detail
 import com.juul.kable.external.GATT_CONN_CANCEL
 import com.juul.kable.external.GATT_CONN_FAIL_ESTABLISH
 import com.juul.kable.external.GATT_CONN_L2C_FAILURE
@@ -37,6 +32,9 @@ import com.juul.kable.gatt.Response.OnDescriptorRead
 import com.juul.kable.gatt.Response.OnDescriptorWrite
 import com.juul.kable.gatt.Response.OnReadRemoteRssi
 import com.juul.kable.gatt.Response.OnServicesDiscovered
+import com.juul.kable.logs.Logger
+import com.juul.kable.logs.Logging
+import com.juul.kable.logs.detail
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.Channel.Factory.CONFLATED
 import kotlinx.coroutines.channels.Channel.Factory.UNLIMITED
@@ -60,7 +58,7 @@ internal class Callback(
     macAddress: String,
 ) : BluetoothGattCallback() {
 
-    private val logger = Logger(logging, tag = "$TAG/Callback", prefix = "$macAddress ")
+    private val logger = Logger(logging, tag = "Kable/Callback", prefix = "$macAddress ")
 
     private var disconnectedAction: DisconnectedAction? = null
     fun invokeOnDisconnected(action: DisconnectedAction) {
@@ -139,7 +137,7 @@ internal class Callback(
             message = "onServicesDiscovered"
             detail(event.status)
         }
-        onResponse.trySendOrLog(event)
+        logger.trySendOrLog(onResponse, event)
     }
 
     override fun onCharacteristicRead(
@@ -155,7 +153,7 @@ internal class Callback(
             detail(event.status)
             detail(value)
         }
-        onResponse.trySendOrLog(event)
+        logger.trySendOrLog(onResponse, event)
     }
 
     override fun onCharacteristicWrite(
@@ -169,7 +167,7 @@ internal class Callback(
             detail(characteristic)
             detail(event.status)
         }
-        onResponse.trySendOrLog(event)
+        logger.trySendOrLog(onResponse, event)
     }
 
     override fun onCharacteristicChanged(
@@ -183,7 +181,7 @@ internal class Callback(
             detail(characteristic)
             detail(value)
         }
-        _onCharacteristicChanged.trySendOrLog(event)
+        logger.trySendOrLog(_onCharacteristicChanged, event)
     }
 
     override fun onDescriptorRead(
@@ -199,7 +197,7 @@ internal class Callback(
             detail(event.status)
             detail(value)
         }
-        onResponse.trySendOrLog(event)
+        logger.trySendOrLog(onResponse, event)
     }
 
     override fun onDescriptorWrite(
@@ -213,7 +211,7 @@ internal class Callback(
             detail(descriptor)
             detail(event.status)
         }
-        onResponse.trySendOrLog(event)
+        logger.trySendOrLog(onResponse, event)
     }
 
     override fun onReliableWriteCompleted(
@@ -238,7 +236,7 @@ internal class Callback(
             detail("rssi", event.rssi)
             detail(event.status)
         }
-        onResponse.trySendOrLog(event)
+        logger.trySendOrLog(onResponse, event)
     }
 
     override fun onMtuChanged(
@@ -252,7 +250,7 @@ internal class Callback(
             detail("mtu", event.mtu)
             detail(event.status)
         }
-        onMtuChanged.trySendOrLog(event)
+        logger.trySendOrLog(onMtuChanged, event)
         if (status == GATT_SUCCESS) this.mtu.value = mtu
     }
 }
@@ -286,8 +284,10 @@ private val Int.connectionStateString: String
         else -> "Unknown($this)"
     }
 
-private fun <E> SendChannel<E>.trySendOrLog(element: E) {
-    trySend(element).getOrElse { cause ->
-        Log.w(TAG, "Callback was unable to deliver $element", cause)
+private fun <E> Logger.trySendOrLog(channel: SendChannel<E>, element: E) {
+    channel.trySend(element).getOrElse { cause ->
+        warn(cause) {
+            message = "Callback was unable to deliver $element"
+        }
     }
 }
