+++ core/src/appleMain/kotlin/PeripheralBuilder.kt
@@ -1,5 +1,7 @@
 package com.juul.kable
 
+import com.juul.kable.logs.Logging
+import com.juul.kable.logs.LoggingBuilder
 import kotlin.coroutines.cancellation.CancellationException
 
 public actual class ServicesDiscoveredPeripheral internal constructor(
