+++ core/src/commonMain/kotlin/logs/SystemLogEngine.kt
@@ -1,0 +1,3 @@
+package com.juul.kable.logs
+
+public expect object SystemLogEngine : LogEngine
