+++ log-engine-tuulbox/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=core
+android.useAndroidX=true
